Terminado endpoint de registro de vuelo

+++ backend/yocos-airlines/vuelo-service/src/main/java/com/pgc/vuelo/repository/AerolineaRepository.java
@@ -2,7 +2,6 @@ package com.pgc.vuelo.repository;
 
 import com.pgc.vuelo.dto.response.AerolineaResponse;
 import com.pgc.vuelo.models.Aerolinea;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -11,11 +10,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AerolineaRepository extends JpaRepository<Aerolinea, Integer> {
+public interface AerolineaRepository extends JpaRepository<Aerolinea, Long> {
 
     @Query("SELECT new com.pgc.vuelo.dto.response.AerolineaResponse(a.id, a.aerolinea) FROM Aerolinea a")
     List<AerolineaResponse> findAllAerolineas();
 
     @Query("SELECT new com.pgc.vuelo.dto.response.AerolineaResponse(a.id, a.aerolinea) FROM Aerolinea a WHERE a.id=:id")
-    AerolineaResponse findAerolineaById(@Param("id") int id);
+    AerolineaResponse findAerolineaById(@Param("id") long id);
 }
